Move dot and batch_dot to a single file
